lol okay interestingness so far!!!
so it is perfect for getting one word rhymes. in can search backward for multiple word rhymes but theyre really silly its hard to get it to pick out things that are good phrases. im using the markov table probabilities but then it just ends up giving a million combinations of "x super common word - y less common word" like for "level" every suggestion ends in "the" lmao. so im still working on it lol